Fix minor thing
* Add log stub
