+++ shell/platform/tizen/BUILD.gn
@@ -37,7 +37,6 @@ _flutter_tizen_source = [
   "flutter_tizen_texture_registrar.cc",
   "key_event_handler.cc",
   "tizen_event_loop.cc",
-  "tizen_log.cc",
   "tizen_renderer.cc",
   "touch_event_handler.cc",
 ]
@@ -131,6 +130,7 @@ template("embedder_for_profile") {
       "channels/settings_channel.cc",
       "external_texture_pixel_gl.cc",
       "external_texture_surface_gl.cc",
+      "tizen_log.cc",
     ]
 
     libs = _libs_minimum
@@ -242,6 +242,7 @@ template("embedder_executable") {
       "channels/localization_channel_stub.cc",
       "channels/platform_channel_stub.cc",
       "channels/settings_channel_stub.cc",
+      "tizen_log_stub.cc",
       "tizen_renderer_evas_gl.cc",
     ]
 
@@ -346,9 +347,7 @@ group("tizen") {
     ":publish_headers_tizen",
   ]
   if (enable_desktop_embeddings) {
-    deps += [
-      ":flutter_tizen_shell",
-    ]
+    deps += [ ":flutter_tizen_shell" ]
   } else {
     deps += [
       ":flutter_tizen_common",
