+++ shell/platform/tizen/tizen_log.h
@@ -8,7 +8,6 @@
 #ifndef __X64_SHELL__
 #include <dlog.h>
 #else
-#include <stdio.h>
 #define log_priority int
 #define DLOG_DEBUG 0
 #define DLOG_WARN 1
@@ -39,9 +38,7 @@ log_priority GetMinLoggingLevel();
 
 #undef __LOG
 
-#ifdef __X64_SHELL__
-#define __LOG(prio, fmt, args...) dlog_print(prio, LOG_TAG, fmt, ##args)
-#elif TV_PROFILE
+#ifdef TV_PROFILE
 // dlog_print() cannot be used because it implicitly passes LOG_ID_APPS as
 // a log id, which is ignored by TV devices. Instead, an internal function
 // __dlog_print() that takes a log id as a parameter is used.
